pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-bot-stack"
version = "0.1.0"
description = "Reusable framework for building Telegram bots with python-telegram-bot"
readme = "README.md"
requires-python = ">=3.9"
authors = [
{name = "telegram-bot-stack contributors", email = "admin@example.com"}
]
license = {text = "MIT"}
keywords = ["telegram", "bot", "framework", "python-telegram-bot", "chatbot"]
classifiers = [
"Development Status :: 3 - Alpha",
"Intended Audience :: Developers",
"Topic :: Communications :: Chat",
"Topic :: Software Development :: Libraries :: Application Frameworks",
"License :: OSI Approved :: MIT License",
"Programming Language :: Python :: 3",
"Programming Language :: Python :: 3.9",
"Programming Language :: Python :: 3.10",
"Programming Language :: Python :: 3.11",
"Programming Language :: Python :: 3.12",
]

dependencies = [
"python-telegram-bot[job-queue]>=22.3,<23.0",
"APScheduler>=3.11.0,<4.0",
"python-dotenv>=1.1.0",
]

[project.optional-dependencies]
# Database support (for future SQL storage backends)
database = [
"sqlalchemy>=2.0,<3.0",
"alembic>=1.13,<2.0",
]
postgres = [
"psycopg2-binary>=2.9",
]

# Full installation with all features
all = [
"sqlalchemy>=2.0,<3.0",
"alembic>=1.13,<2.0",
"psycopg2-binary>=2.9",
]

# Development dependencies
dev = [
"ruff>=0.8.0",
"mypy>=1.17.0",
"pytest>=8.4.0",
"pytest-asyncio>=0.24.0",
"pytest-cov>=4.1.0",
"pytest-mock>=3.12.0",
"pre-commit>=3.8.0",
]

# GitHub Actions automation
github-actions = [
"PyGithub>=2.1.1",
"openai>=1.3.0",
"python-dotenv>=1.0.0",
"PyYAML>=6.0",
"gitpython>=3.1.40",
]

[project.urls]
Homepage = "https://github.com/sensiloles/telegram-bot-stack"
Repository = "https://github.com/sensiloles/telegram-bot-stack"
Documentation = "https://github.com/sensiloles/telegram-bot-stack#readme"
"Issue Tracker" = "https://github.com/sensiloles/telegram-bot-stack/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["telegram_bot_stack*"]
exclude = ["tests*", "scripts*", "examples*", "src*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
exclude = [
"venv/",
"build/",
"dist/"
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
# Start with basic rules, gradually adding stricter ones
select = [
"E",             # pycodestyle errors
"F",             # pyflakes
"W",             # pycodestyle warnings
"I",             # isort (import sorting)
"UP",            # pyupgrade (syntax modernization)
"B",             # flake8-bugbear (likely bugs)
]

# Ignore rules for gradual adoption
ignore = [
# Type annotations - will add later
"ANN",
# Logging format - too strict for initial setup
"G004",
# Exception handling - will fix gradually
"BLE001", "TRY400", "TRY300", "B904",
# Unused arguments - many needed for API compatibility
"ARG002", "ARG001",
# Unused variables - will fix gradually
"F841",
# Magic values - will fix later
"PLR2004",
# Path operations - will modernize later
"PTH123",
# Datetime timezone - will fix later
"DTZ001",
# Complex rules for later adoption
"RUF005", "RUF001", "SLF001",
# Formatter conflicts
"COM812", "ISC001",
# Import organization - will fix later
"E402",
# Modern typing - will fix later
"UP035", "UP006",
# Long lines - formatter will handle
"E501"
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["T201"]            # Allow print statements in scripts
"tests/*" = ["ANN", "PLR2004"]    # Relax type annotations and magic values in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
"--cov=telegram_bot_stack",
"--cov-report=html",
"--cov-report=term-missing",
"--cov-report=xml",
"--strict-markers",
"-v",
]
markers = [
"asyncio: mark test as async",
"integration: mark test as integration test",
"unit: mark test as unit test",
]
filterwarnings = [
"ignore::DeprecationWarning",
"ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["telegram_bot_stack"]
omit = [
"*/tests/*",
"*/venv/*",
"*/__pycache__/*",
"*/site-packages/*",
"*/examples/*",
]

[tool.coverage.report]
exclude_lines = [
"pragma: no cover",
"def __repr__",
"raise AssertionError",
"raise NotImplementedError",
"if __name__ == .__main__.:",
"if TYPE_CHECKING:",
"@abstractmethod",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "$version"
version_files = [
"pyproject.toml:version",
"telegram_bot_stack/__init__.py:__version__"
]
